modules/cell-sensor/requirements.txt
viam-sdk==0.69.0
typing_extensions>=4.0.0
httpx>=0.27.0
